Change int to numeric for time-range

+++ db/migrate/20250124193727_add_time_range_to_conveyance.rb
@@ -1,6 +1,6 @@
 class AddTimeRangeToConveyance < ActiveRecord::Migration[7.2]
   def change
-    add_column :conveyances, :start_time, :integer
-    add_column :conveyances, :end_time, :integer
+    add_column :conveyances, :start_time, :numeric
+    add_column :conveyances, :end_time, :numeric
   end
 end
